Extends ARX

+++ Util/tools.py
@@ -49,7 +49,7 @@ def mse(y_true: np.ndarray, y_pred: np.ndarray) -> float:
     mse_ : float
         Mean squared error.
     """
-    mse_ = np.mean((y_true - y_pred) ** 2)
+    mse_ = float(np.mean((y_true - y_pred) ** 2))
     return mse_
 
 
@@ -69,7 +69,7 @@ def mae(y_true: np.ndarray, y_pred: np.ndarray) -> float:
     mae_ : float
         Mean absolute error.
     """
-    mae_ = np.mean(np.abs(y_true - y_pred))
+    mae_ = float(np.mean(np.abs(y_true - y_pred)))
     return mae_
 
 
@@ -93,7 +93,7 @@ def r2(y_true: np.ndarray, y_pred: np.ndarray) -> float:
     if ss_total == 0:
         return 0.0
     ss_residual = np.sum((y_true - y_pred) ** 2)
-    r2_ = 1 - (ss_residual / ss_total)
+    r2_ = float(1 - (ss_residual / ss_total))
     return r2_
 
 
@@ -119,7 +119,7 @@ def r2_adjusted(y_true: np.ndarray, y_pred: np.ndarray, n_features: int) -> floa
     if m <= n_features + 1:
         return np.nan
     r2_ = r2(y_true, y_pred)
-    r2_adj = 1 - (1 - r2_) * (m - 1) / (m - n_features - 1)
+    r2_adj = float(1 - (1 - r2_) * (m - 1) / (m - n_features - 1))
     return r2_adj
 
 
@@ -139,7 +139,7 @@ def smape(y_true: np.ndarray, y_pred: np.ndarray) -> float:
     smape_ : float
         SMAPE score, expressed as a percentage.
     """
-    smape_ = 100 * np.mean(2 * np.abs(y_true - y_pred) / (np.abs(y_true) + np.abs(y_pred) + 1e-10))
+    smape_ = float(100 * np.mean(2 * np.abs(y_true - y_pred) / (np.abs(y_true) + np.abs(y_pred) + 1e-10)))
     return smape_
 
 
@@ -163,7 +163,7 @@ def diagnose(y_true: np.ndarray, y_pred: np.ndarray, n_features: int = 1) -> dic
     metrics : dict
         Dictionary containing MSE, MAE, R^2, Adjusted R^2, SMAPE, and m.
     """
-    
+
     return {
         "MSE"   : mse(y_true, y_pred),
         "MAE"   : mae(y_true, y_pred),
